+++ .github/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 (Dates are in YYYY-MM-DD format. This message is mainly for my own sake.)
 
 ## [Unreleased]
+### Added
+* Pronouns are now automatically detected in verification messages and the relevant roles applied on verification acceptance.
 
 ## [3.6.2] - 2021-04-08
 ### Changed